Refactor page.svelte
